services/ai/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
httpx>=0.25.0
openai>=1.3.0
google-generativeai>=0.3.0
groq>=0.4.0
vercel-ai-sdk>=0.1.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
tenacity>=8.2.3